Added Basic Priest Functionality + Rearrange
-Created middle-man script "AuraVillager" that Mage and Priest inherit from as they both use auras
-Added Placeholder Priest Aura
-Cleaned scripts up
-Added middle-man script "Aura for that Priest Aura and Mage Aura inherit from
-Added logic for Priest Aura so it protected players
-Imporvements to aura collisions

+++ Assets/Scripts/Boss/Head.cs
@@ -54,7 +54,7 @@ public class Head : MonoBehaviour {
         }
     }
 
-    public void OnHit(int damageMultiplier)
+    public void OnHit(float damageMultiplier)
     {
         Debug.Log("Golem took " + 1 * damageMultiplier + " Damage!");
         Golem.health -= 1 * damageMultiplier;
